+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   - New function `delete` to delete a list
   - New function `remove_member` to remove a member from a list
   - New function `subscribe` to subscribe to a list
+  - New function `unsubscribe` to unsubscribe from a list
 - New enum `service::ListMethod` and `list` field in `service::RateLimitStatus` to contain the rate
   limit info for these methods